requirement.txt
absl-py==1.0.0
albumentations==1.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cached-property==1.5.2
cffi==1.15.0
charset-normalizer==2.0.12
clang==5.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
dlib @ file:///D:/bld/dlib_1652035402968/work
dominate==2.6.0
dpcpp-cpp-rt==2022.0.3
easydict==1.9
einops==0.4.1
entrypoints==0.4
fastjsonschema==2.15.3
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.2.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.2
h5py==2.10.0
image==1.5.33
imageio==2.19.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
imutils==0.5.4
intel-cmplr-lib-rt==2022.0.3
intel-cmplr-lic-rt==2022.0.3
intel-opencl-rt==2022.0.3
ipykernel==6.13.0
ipython==7.33.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso8601==1.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl==2022.0.3
mkl-fft==1.2.0
mkl-random==1.2.1
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.11
numpy @ file:///D:/bld/numpy_1649806536290/work
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyserial==3.5
PySimpleGUI==4.60.0
python-dateutil==2.8.2
python-xz==0.4.0
pytz==2022.1
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
qudida==0.0.4
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
serial==0.0.97
six==1.16.0
soupsieve==2.3.2.post1
tbb4py==2021.5.2
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboardX==2.5
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
urllib3==1.26.9
vc==2018.7.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.12.1
yellowbrick==1.4
zipp==3.8.0
